+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - The `score_functions.disabled()` function, used to force a `Mapper` to work in override-only mode.
+- An optional argument `for_value` to the `heuristic_functions.value_fstring_alias()` function.
 
 ### Changed
 - Add `n_splits` argument to `TimeFold` (scikit-learn compatibility).
@@ -19,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - The `heuristic_functions.candidate_fstring_alias()` function may now properly use _value_ and _context_ placeholders.
+- Properly raise exceptions in `[value/candidate]_fstring_alias` for invalid `fstring` arguments.
 
 ## [2.1.0] - 2023-02-11